cardService now has the allCards ssot

+++ services/Card.ts
@@ -3,18 +3,47 @@ import { Card, CardCreate, CardUpdate } from "@/types/Card";
 import { KnowledgeLevel } from "@/types/KnowledgeLevel";
 import { Repository } from "typeorm";
 
+export type CardChangeListener = () => void;
+
 export class CardService {
   readonly RELATIONS = ["tags"];
+  private listeners: CardChangeListener[] = [];
+
+  _allCards: Card[] = [];
 
   constructor(private repo: Repository<CardSchema>) {}
+
+  onChange(listener: CardChangeListener) {
+    this.listeners.push(listener);
+  }
+
+  offChange(listener: CardChangeListener) {
+    console.log("offChange");
+    this.listeners = this.listeners.filter((l) => l !== listener);
+  }
+
+  private notifyChange() {
+    this.listeners.forEach((listener) => listener());
+  }
+
+  async init() {
+    this._allCards = await this.getAll();
+  }
 
   // TODO add tag in ctor
   async create(payload: CardCreate): Promise<Card> {
     const entity = this.repo.create();
 
+    // TODO validate input is a valid KnowledgeLevel
     Object.assign(entity, { ...payload });
 
-    return (await this.repo.save(entity)) as Card;
+    const ret = (await this.repo.save(entity)) as Card;
+
+    this._allCards.push(ret);
+
+    this.notifyChange();
+
+    return ret;
   }
 
   async getById(id: number): Promise<Card> {
@@ -36,12 +65,40 @@ export class CardService {
       return;
     }
 
+    // TODO validate input is a valid KnowledgeLevel
     Object.assign(entity, { ...payload });
 
-    return (await this.repo.save(entity)) as Card;
+    const ret = (await this.repo.save(entity)) as Card;
+
+    const index = this._allCards.findIndex((card) => card.id === ret.id);
+    if (index !== -1) {
+      this._allCards[index] = ret;
+    }
+
+    this.notifyChange();
+
+    return ret;
   }
 
   async delete(id: CardSchema["id"]) {
+    const entity = await this.repo.findOne({ where: { id } });
+    if (!entity) {
+      // TODO think of error handling
+      console.error(`Card with id ${id} not found`);
+      return;
+    }
+
+    const index = this._allCards.findIndex((card) => card.id === entity.id);
+    if (index !== -1) {
+      this._allCards.splice(index, 1);
+    }
+
+    this.notifyChange();
+
     await this.repo.delete({ id });
+  }
+
+  get allCards() {
+    return this._allCards;
   }
 }
